# API/crystal.api.1.0.0-swagger2.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.01",
    "title": "Crystal Server API",
    "description": "This is a simple Crystal server API. You can find out more about Crystal at <a href='http://getcrystal.net'>http://getcrystal.net</a>  For this sample, you can use the api key 'get-crystal' for tests that need authorization (authorization filters)",
    "license": {
      "name": "MIT",
      "url": "http://github.com/gruntjs/grunt/blob/master/LICENSE-MIT"
    }
  },
  "host": "crystal.getcrystal.net",
  "basePath": "/api/01",
  "securityDefinitions": {},
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/agent": {
      "post": {
        "description": "Create agent",
        "tags": [
          "Agent"
        ],
        "operationId": "createAgent_",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": false,
            "x-is-map": false,
            "schema": {
              "$ref": "#/definitions/Agent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/agent/createWithList": {
      "post": {
        "description": "Creates list of agents with given list input",
        "tags": [
          "Agent"
        ],
        "operationId": "createAgentsWithListInput_",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": false,
            "x-is-map": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Agent"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/agent/{agentId}": {
      "get": {
        "description": "Find agent by ID",
        "tags": [
          "Agent"
        ],
        "operationId": "getAgentById_",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "string",
            "description": "hash ID of agent that needs to be fetched"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Agent"
            }
          },
          "400": {
            "description": "Error 400"
          },
          "404": {
            "description": "Error 404"
          }
        }
      },
      "post": {
        "description": "Updates an agent by post of form data",
        "tags": [
          "Agent"
        ],
        "operationId": "updateAgentWithForm_",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "string",
            "description": "hash ID of agent that needs to be fetched"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object"
            }
          },
          "405": {
            "description": "Error 405"
          }
        }
      },
      "delete": {
        "description": "Deletes an agent",
        "tags": [
          "Agent"
        ],
        "operationId": "deleteAgent_",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "string",
            "description": "hash ID of agent that needs to be fetched"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error 400"
          }
        }
      }
    },
    "/agent/{agentId}/roles": {
      "get": {
        "description": "get agent roles by agent ID",
        "tags": [
          "Agent"
        ],
        "operationId": "getAgentRolesByAgentId_",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Role"
              }
            }
          },
          "400": {
            "description": "Error 400"
          }
        }
      }
    },
    "/agent/{agentlogin}": {
      "get": {
        "description": "Get agent by login name",
        "tags": [
          "Agent"
        ],
        "operationId": "getAgentByName_",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "agentlogin",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Agent"
            }
          },
          "400": {
            "description": "Error 400"
          },
          "404": {
            "description": "Error 404"
          }
        }
      },
      "put": {
        "description": "Updated agent",
        "tags": [
          "Agent"
        ],
        "operationId": "updateUser_",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "agentlogin",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "string",
            "description": "name that need to be deleted"
          },
          {
            "name": "body",
            "in": "body",
            "required": false,
            "x-is-map": false,
            "schema": {
              "$ref": "#/definitions/Agent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "Error 400"
          },
          "404": {
            "description": "Error 404"
          }
        }
      },
      "delete": {
        "description": "Delete agent by user login",
        "tags": [
          "Agent"
        ],
        "operationId": "deleteUser_",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "agentlogin",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "string",
            "description": "name that need to be deleted"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error 400"
          },
          "404": {
            "description": "Error 404"
          }
        }
      }
    },
...
    "/contest": {
      "post": {
        "description": "create contest",
        "tags": [
          "Contest"
        ],
        "operationId": "createContest_",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "x-is-map": false,
            "description": "body as contest object",
            "schema": {
              "$ref": "#/definitions/Contest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
...
    "/agent/{reputationId}/{min}/{max}": {
      "get": {
        "description": "GetAgentsByReputation",
        "tags": [
          "Reputation"
        ],
        "operationId": "getAgentsByReputationRange_",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "reputationId",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "string"
          },
          {
            "name": "min",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "max",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Agent"
              }
            }
          },
          "400": {
            "description": "Error 400"
          }
        }
      }
    },
...
    "/contest/{communityid}": {
      "get": {
        "description": "get all contests for a community",
        "tags": [
          "Contest"
        ],
        "operationId": "getContestsByCommunity_",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "communityid",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Contest"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "Agent": {
      "title": "Agent",
      "description": "Agent currently represents only users (persons that create logins). In the future will also represent code-agents (contracts, bots, etc.)",
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "lastName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "userStatus": {
          "description": "User Status - active, inactive, banned",
          "type": "integer",
          "format": "int32"
        },
        "userlogin": {
          "type": "string"
        }
      },
      "required": [
        "id"
      ]
    },
...
    "Role": {
      "title": "Role",
      "description": "a Role represents permissions to a UNC path for CRUD and Execute operations.",
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "path": {
          "description": "rest endpoint or registered contract endpoint (allows for regex wildcards) to resource(s) that this role can access or execute",
          "type": "string"
        },
        "permissions": {
          "description": "list of permissions for the resources in the path. Agents assigned this role will be allowed to perform GET/POST/DELETE/PATCH (read/update/delete) on the resource.",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Permission"
          }
        }
      },
      "required": [
        "id",
        "name",
        "path",
        "permissions"
      ]
    },
    "Permission": {
      "title": "Permission",
      "description": "a permission for read, write or execute on some unique URI that represents a function or resource. These resources will be the components of the governance scheme, e.g. they will be data or functions that in aggregate comprise a particular governance model.",
      "type": "object",
      "properties": {
        "flags": {
          "description": "create, read, write, execute / or GET/POST",
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        }
      },
      "required": [
        "flags",
        "id",
        "name"
      ]
    },
    "Content": {
      "title": "Content",
      "description": "the content for output based contests",
      "type": "object",
      "properties": {
        "contentName": {
          "type": "string"
        },
        "explainations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Explanation"
          }
        },
        "id": {
          "type": "string"
        },
        "ratings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Rating"
          }
        },
        "submitter": {
          "description": "the creator or owner of the content",
          "$ref": "#/definitions/Agent"
        }
      },
      "required": [
        "contentName",
        "id",
        "submitter"
      ]
    },
    "Explanation": {
      "title": "Explanation",
      "description": "content rating as text",
      "type": "object",
      "properties": {
        "contentId": {
          "description": "link to content being explained",
          "type": "string"
        },
        "description": {
          "description": "explanation description",
          "type": "string"
        },
        "explainedBy": {
          "description": "explainer id",
          "type": "string"
        },
        "name": {
          "description": "explanation name",
          "type": "string"
        }
      },
      "required": [
        "contentId",
        "description",
        "name"
      ]
    },
    "Rating": {
      "title": "Rating",
      "type": "object",
      "properties": {
        "contentId": {
          "description": "the content being rated",
          "type": "string"
        },
        "type": {
          "description": "rating type",
          "type": "string"
        },
        "value": {
          "description": "rating value TODO: what is data type - string, number?",
          "type": "string"
        }
      },
      "required": [
        "contentId",
        "type",
        "value"
      ]
    },
    "Contest": {
      "title": "Contest",
      "type": "object",
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Content"
          }
        },
        "contestType": {
          "description": "the types of contests",
          "$ref": "#/definitions/ContestType"
        },
        "creator": {
          "description": "contest creator",
          "$ref": "#/definitions/Agent"
        },
        "description": {
          "type": "string"
        },
        "id": {
          "description": "hash for this contest",
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "startDate": {
          "type": "string",
          "format": "date-time"
        },
        "endDate": {
          "type": "string",
          "format": "date-time"
        },
        "communityId": {
          "description": "community this contest belongs to",
          "type": "string"
        }
      },
      "required": [
        "contestType",
        "creator",
        "description",
        "id",
        "name",
        "communityId"
      ]
    },
    "ContestType": {
      "title": "ContestType",
      "description": "the types of contests",
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      },
      "required": [
        "description",
        "id",
        "name"
      ]
    },
...
  }
}